ゲスト対応

+++ src/main/java/co/jp/wever/graphql/infrastructure/converter/entity/section/SectionNumberEntityConverter.java
@@ -8,7 +8,7 @@ public class SectionNumberEntityConverter {
     public static SectionNumberEntity toSectionNumberEntity(Map<String, Object> resultMap) {
         return SectionNumberEntity.builder()
                                   .id((String) resultMap.get("id"))
-                                  .number((int) resultMap.get("number"))
+                                  .number((long) resultMap.get("number"))
                                   .build();
     }
 }
